chore: move startup and teardown in support folder
