+++ e2e/knative/support/teardown_test.go
@@ -20,7 +20,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package knative
+package support
 
 import (
 	"testing"
